refactor: decompose KuzuGraphBackend (728→247+285+93) into partial classes
Split by responsibility: shell+schema, CRUD operations, community detection.
